Renamed and merged conditions into CondHasBuildRestrictions. Also changed EffBuildRestrictions' doc annotations.
